refactoring dialog card, dialog utils, and many style changes
